refactor and javadocs
mosly added javadocs and fixed some checkstyle, small refactoring
