+++ src/main/java/com/funkydonkies/gamestates/BallState.java
@@ -1,6 +1,7 @@
 package com.funkydonkies.gamestates;
 
-import com.funkydonkies.controllers.SuperSizePowerup;
+import powerups.SuperSizePowerup;
+
 import com.funkydonkies.exceptions.BadDynamicTypeException;
 import com.funkydonkies.w4v3.App;
 import com.funkydonkies.w4v3.Ball;
